requirements.txt
# Python dependencies for TILL MS Access to SQL Migration

# Database connectivity
pyodbc>=5.0.0
pymssql>=2.2.0
sqlalchemy>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Data analysis and manipulation
pandas>=2.0.0
numpy>=1.24.0

# Excel export (optional)
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Reporting and visualization
tabulate>=0.9.0
colorama>=0.4.6

# For Access database handling (Windows only)
pywin32>=306; platform_system=="Windows"

# Date/time utilities
python-dateutil>=2.8.0

# JSON handling
simplejson>=3.19.0
